refactor: split app context module
